pyproject.toml
[project]
name = "ai-style-generator"
version = "0.1.0"
description = ""
authors = [
    {name = "RDC28",email = "rohitrdc12345@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai (>=1.33.0,<2.0.0)"
]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
flask = "^3.1.2"
flask-cors = "^6.0.1"
python-dotenv = "^1.1.1"
requests = "^2.32.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
